requirements.txt
# After installing all of the below, run:
# python -m spacy download en_core_web_sm

# IBM Watson & AI Platform
ibm-watson>=6.0.0
ibm-cloud-sdk-core>=3.16.0

# Web Framework & API
fastapi>=0.100.0
uvicorn[standard]>=0.22.0
streamlit>=1.25.0

# Document Processing
PyPDF2>=3.0.1
pdfplumber>=0.9.0
python-docx>=0.8.11
reportlab>=4.0.4

# Natural Language Processing & Machine Learning
spacy>=3.6.0
sentence-transformers>=2.2.2
scikit-learn>=1.3.0
pandas>=2.0.0
numpy>=1.24.0

# Text Processing & Analysis
nltk>=3.8.1
textstat>=0.7.3

# Utilities & File Handling
requests>=2.31.0
beautifulsoup4>=4.12.0
python-multipart>=0.0.6
aiofiles>=23.0.0

# Database (if needed)
sqlalchemy>=2.0.0

# Development & Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.7.0
flake8>=6.0.0

# Environment & Configuration
python-dotenv>=1.0.0
pydantic>=2.0.0